fix: regnerate templates
