+++ templates/ffcli/grpc/templates/cmd/%7B%7B.File.Options.GoPackage%7Creplace%20%22%2E%2F%3B%22%20%22%22%7D%7D/main.go.tmpl
