+++ templates/ffcli/unix-cmd/pkg/swissknife_gen/z_swissknife_gen.go
@@ -1,47 +1,49 @@
 package swissknife_gen
 
 import (
-	"context"
-	"log"
+    "context"
+    "log"
 
-	"google.golang.org/grpc"
-	"moul.io/adapterkit/pkg/lib"
+    "google.golang.org/grpc"
+    "moul.io/adapterkit/pkg/lib"
 
-	swissknife "github.com/pmg-tools/adapterkit-module-swissknife"
+    swissknife "$[ADAPTERKIT_LOGIC_PACKAGE]"
 )
+
 
 func SvcConvHexa(input string, svc swissknife.SwissknifeSvcServer) (string, error) {
 	ctx := context.Background()
 
 	conn, err := grpc.DialContext(ctx, "",
 		grpc.WithInsecure(),
-		grpc.WithContextDialer(lib.Dialer(svc, swissknife.RegisterSwissknifeSvcServer)))
-	if err != nil {
+ 		grpc.WithContextDialer(lib.Dialer(svc, swissknife.RegisterSwissknifeSvcServer)))
+    if err != nil {
 		log.Fatal(err)
 	}
-	defer conn.Close()
-
-	client := swissknife.NewSwissknifeSvcClient(conn)
-	req := &swissknife.ConvHexaReq{Input: input}
-	res, err := client.ConvHexa(ctx, req)
-	return res.Output, err
+    defer conn.Close()
+    
+    client := swissknife.NewSwissknifeSvcClient(conn)
+    req := &swissknife.ConvHexaReq{Input: input}
+    res, err := client.ConvHexa(ctx, req)
+    return res.Output, err
 }
+
 
 func SvcConvBase64(input string, svc swissknife.SwissknifeSvcServer) (string, error) {
 	ctx := context.Background()
 
 	conn, err := grpc.DialContext(ctx, "",
 		grpc.WithInsecure(),
-		grpc.WithContextDialer(lib.Dialer(svc, swissknife.RegisterSwissknifeSvcServer)))
-	if err != nil {
+ 		grpc.WithContextDialer(lib.Dialer(svc, swissknife.RegisterSwissknifeSvcServer)))
+    if err != nil {
 		log.Fatal(err)
 	}
-	defer conn.Close()
-
-	client := swissknife.NewSwissknifeSvcClient(conn)
-	req := &swissknife.ConvBase64Req{Input: input}
-	res, err := client.ConvBase64(ctx, req)
-	return res.Output, err
+    defer conn.Close()
+    
+    client := swissknife.NewSwissknifeSvcClient(conn)
+    req := &swissknife.ConvBase64Req{Input: input}
+    res, err := client.ConvBase64(ctx, req)
+    return res.Output, err
 }
